requirements.txt
aiogram
pydantic
python-dotenv
Jinja2
aiohttp-client-cache[sqlite]
